исправил ошибку с вводов

+++ PracticKotlin5/src/io/io.kt
@@ -10,19 +10,19 @@ class InputString : Read {
         return readLine()
     }
 }
-//TODO ПРИ ВВОДЕ ПЛОХО ОБРАБАТЫВАЕТСЯ CTRL+D
+
 class LooperInput(
     private val stringValidation : StringValidation = ValidateString(),
     private val outputMessage : PrintMessage = Speaker(),
     private val readString : Read = InputString()
 ) : InputLooper {
-    override fun input(message: String, regex: Regex): String {
-        var inputString: String? = null
-        while (!stringValidation.check(regex, inputString)) {
+    override fun input(message: String, regex: Regex): String? {
+        var inputString: String? = ""
+        while (!stringValidation.check(regex, inputString) && inputString != null) {
             outputMessage.printMessage(message)
             inputString = readString.readString()
         }
-        return inputString!!
+        return inputString
     }
 }
 
@@ -32,19 +32,23 @@ class InputIndex(
     private val stringValidation : StringValidation = ValidateString(),
     private val inputString : Read = InputString()
 ) : RowIndex {
-    override fun getIndex(message: String, listSize : Int): Int {
-        var indexString : String?
-        var indexInt = -1
+    override fun getIndex(message: String, listSize : Int): Int? {
+        var indexInt : Int? = -1
         while (indexInt == -1) {
             outputMessage.printMessage(message)
-            indexString = inputString.readString()
-            if (stringValidation.check("^[1-9]+\\d*".toRegex(), indexString)) {
-                indexInt = intParser.parseToInt(indexString!!)
-                if (indexInt !in 1..listSize)
-                    indexInt = -1
+            val indexString: String? = inputString.readString()
+            if (indexString != null) {
+                if (stringValidation.check("^[1-9]+\\d*".toRegex(), indexString)) {
+                    indexInt = intParser.parseToInt(indexString)
+                    if (indexInt !in 1..listSize)
+                        indexInt = -1
+                } else
+                    outputMessage.printMessage("\nНеправильный номер строки, попробуйте снова\n")
             }
-            else
-                outputMessage.printMessage("\nНеправильный номер строки, попробуйте снова\n")
+            else {
+                outputMessage.printMessage("\nВыход из программы\n")
+                indexInt = null
+            }
         }
         return indexInt
     }
@@ -59,7 +63,10 @@ class DeleteRow(
         if (dataBase.getList().toList().isNotEmpty()) {
             outputData.execute(dataBase)
             val index = inputIndex.getIndex("Введите строку, которую хотите удалить -> ", dataBase.getList().toList().size)
-            dataBase.delete(index - 1)
+            if (index != null)
+                dataBase.delete(index - 1)
+            else
+                outputMessage.printMessage("\nВыход из программы\n")
         }
         else {
             outputMessage.printMessage("\nПустая база данных\n")
@@ -75,7 +82,11 @@ class InputDataRow(
         listColumns.forEach {
             it.add(nullableDataRow)
         }
-        dataBase.add(nullableDataRow.asDataRow())
+        if (nullableDataRow.product != null &&
+            nullableDataRow.price != null &&
+            nullableDataRow.count != null &&
+            nullableDataRow.dateOfBuy != null)
+            dataBase.add(nullableDataRow.asDataRow())
     }
 }
 
@@ -91,12 +102,22 @@ class EditRow(
             outputData.execute(dataBase)
             val indexRow = inputIndex.getIndex("Введите номер строки, которую хотите редактировать -> ",
                                                 dataBase.getList().toList().size)
-            outputColumns.printColumns(listColumns.map { it.name })
-            val indexColumns = inputIndex.getIndex("Введите номер столбца, который хотите редактировать -> ",
-                                                    listColumns.size)
-            val currentRow = dataBase.getList().toList()[indexRow - 1]
-            listColumns.getOrNull(indexColumns - 1)?.edit(currentRow)
-            dataBase.set(indexRow - 1, currentRow)
+            if (indexRow != null) {
+                outputColumns.printColumns(listColumns.map { it.name })
+                val indexColumns = inputIndex.getIndex(
+                    "Введите номер столбца, который хотите редактировать -> ",
+                    listColumns.size
+                )
+                if (indexColumns != null) {
+                    val currentRow = dataBase.getList().toList()[indexRow - 1]
+                    listColumns.getOrNull(indexColumns - 1)?.edit(currentRow)
+                    dataBase.set(indexRow - 1, currentRow)
+                }
+                else
+                    outputMessage.printMessage("\nВыход из программы\n")
+            }
+            else
+                outputMessage.printMessage("\nВыход из программы\n")
         }
         else
             outputMessage.printMessage("\nDataBase is empty!")
@@ -107,6 +128,7 @@ class SortTable(
     private val inputIndex : RowIndex = InputIndex(),
     private val printer: Printer = Printer(),
     private val outputColumns : PrintColumns = OutputColumns(),
+    private val outputMessage: PrintMessage = Speaker(),
     private val listColumns : List<Column> = listOf(ProductColumn(), PriceColumn(), CountColumn(), DateColumn())
 ) : ExecutorCommand {
     override fun execute(dataBase: Table) {
@@ -115,10 +137,14 @@ class SortTable(
             "Введите номер столбца, по которому хотите сортировать -> ",
             listColumns.size
         )
-        val list = dataBase.getList().toList()
-        printer.show(
-            listColumns[indexColumn - 1].sort(list)
-        )
+        if (indexColumn != null) {
+            val list = dataBase.getList().toList()
+            printer.show(
+                listColumns[indexColumn - 1].sort(list)
+            )
+        }
+        else
+            outputMessage.printMessage("\nВыход из программы\n")
     }
 }
 
@@ -126,6 +152,7 @@ class SearchRow(
     private val inputIndex : RowIndex = InputIndex(),
     private val printer: Printer = Printer(),
     private val outputColumns : PrintColumns = OutputColumns(),
+    private val outputMessage: PrintMessage = Speaker(),
     private val listColumns : List<Column> = listOf(ProductColumn(), PriceColumn(), CountColumn(), DateColumn())
 ) : ExecutorCommand {
     override fun execute(dataBase: Table) {
@@ -134,10 +161,14 @@ class SearchRow(
             "Введите столбца, по которому хотите найти -> ",
             listColumns.size
         )
-        val list = dataBase.getList().toList()
-        printer.show(
-            listColumns[indexColumn - 1].search(list)
-        )
+        if (indexColumn != null) {
+            val list = dataBase.getList().toList()
+            printer.show(
+                listColumns[indexColumn - 1].search(list)
+            )
+        }
+        else
+            outputMessage.printMessage("\nВыход из программы\n")
     }
 }
